12/15 사용자 권한관리 개발 (저장 보완 필요)

+++ src/main/java/com/lil/pretty/domain/admin/user/controller/UserMainController.java
@@ -5,16 +5,23 @@ import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.lil.pretty.config.security.auth.AuthDetail;
+import com.lil.pretty.config.security.auth.AuthDetailService;
+import com.lil.pretty.config.security.jwt.JwtTokenProvider;
 import com.lil.pretty.domain.admin.user.dto.UserMain;
 import com.lil.pretty.domain.admin.user.service.UserMainService;
+import com.lil.pretty.domain.common.dto.CUDCommonResponse;
 import com.lil.pretty.domain.common.dto.CommonResponse;
 import com.lil.pretty.domain.user.model.User;
+import com.lil.pretty.domain.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -22,14 +29,14 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("api/admin/user/main")
 @RequiredArgsConstructor
 public class UserMainController {
-
 	private final UserMainService userMainService;
+
+	//사용자 조회
 	@PostMapping("/getAdminUserItems")
 	public ResponseEntity<CommonResponse> getAdminUserItems(@RequestBody Map<String, String> request){
 	
 		try{
 			String usId = request.get("usId");
-			System.out.print("민정 테스트 usid"+usId);
 			List<Map<String,Object>> items = userMainService.getAdminUserItems(usId);
 			return ResponseEntity.ok(new CommonResponse(true, items, "Success"));
 		}
@@ -39,29 +46,45 @@ public class UserMainController {
 	}
 	//사용자 삭제
 	@PostMapping("/deleteAdminUserItems")
-	public ResponseEntity<CommonResponse> deleteAdminUserItems(@RequestBody Map<String, String> request){
-	
+	public ResponseEntity<CUDCommonResponse> deleteAdminUserItems(@RequestBody Map<String, String> request){
+		
+		String usId = request.get("usId");
 		try{
-			String usId = request.get("usId");
-			
-			int items = userMainService.deleteAdminUserItems(usId);
-			return ResponseEntity.ok(new CommonResponse(true, items, "Success"));
+			if (usId == null || usId.isEmpty()) {
+                return ResponseEntity.badRequest().body(new CUDCommonResponse<>(0, 0, null, "삭제할 대상이 없습니다."));
+            }
+			CUDCommonResponse result = userMainService.deleteAdminUserItems(usId);
+
+			return ResponseEntity.ok(new CUDCommonResponse<>(0, 0, result, "Success"));
+
 		}
 		catch(Exception e) {
-			return ResponseEntity.status(401).body(new CommonResponse(false, null,"Failed"));
+			 return ResponseEntity.status(500).body(new CUDCommonResponse<>(0, 0, null, "삭제 처리 중 오류가 발생했습니다."));
 		}
 	}
 	//사용자 수정
 	@PostMapping("/updateAdminUserItems")
-	public ResponseEntity<CommonResponse> updateAdminUserItems(@RequestBody UserMain userRequest){
+	public ResponseEntity<CUDCommonResponse> updateAdminUserItems(@RequestBody UserMain userRequest){
 	
 		try{
 			
-	        int items = userMainService.updateAdminUserItems(userRequest);
-	        return ResponseEntity.ok(new CommonResponse(true, items, "Success"));
+			CUDCommonResponse result = userMainService.updateAdminUserItems(userRequest);
+			return ResponseEntity.ok(new CUDCommonResponse<>(0, 0, result, "Success"));
 		}
 		catch(Exception e) {
-			return ResponseEntity.status(401).body(new CommonResponse(false, null,"Failed"));
+			return ResponseEntity.status(500).body(new CUDCommonResponse<>(0, 0, null, "수정 처리 중 오류가 발생했습니다."));
 		}
 	}
+	//사용자 추가
+		@PostMapping("/insertAdminUserItems")
+		public ResponseEntity<CUDCommonResponse> insertAdminUserItems(@RequestBody UserMain userRequest, @AuthenticationPrincipal(expression = "username") String userId){
+		
+			try{
+				CUDCommonResponse result = userMainService.insertAdminUserItems(userRequest,userId);
+				return ResponseEntity.ok(new CUDCommonResponse<>(0, 0, result, "Success"));
+			}
+			catch(Exception e) {
+				return ResponseEntity.status(500).body(new CUDCommonResponse<>(0, 0, null, "저장 처리 중 오류가 발생했습니다."));
+			}
+		}	
 }
